Added test for deleting reports

+++ reports/brokers/api/views/deleting_old_reports.py
@@ -19,23 +19,14 @@ class ReportCleaner(Greenlet):
 
     def deleting_old_reports(self):
         for file in os.listdir(self.result_dir):
-            file_date = datetime.strptime(str(file.split('_report-number=')[0]), '%Y-%m-%d-%H-%M-%S')
-            now = datetime.now()
-            delta = now - file_date
-            if delta.days >= 1:
-                os.remove(os.path.abspath(os.path.join(self.result_dir, file)))
+            if os.path.splitext(file)[1] == '.xlsx':
+                file_date = datetime.strptime(str(file.split('_report-number=')[0]), '%Y-%m-%d-%H-%M-%S')
+                now = datetime.now()
+                delta = now - file_date
+                if delta.days >= 1:
+                    os.remove(os.path.abspath(os.path.join(self.result_dir, file)))
 
-    def launch(self):
+    def _run(self):
         while True:
             self.deleting_old_reports()
-            print('.')
             gevent.sleep(DELAY)
-
-
-def main():
-    report_cleaner = ReportCleaner()
-    report_cleaner.launch()
-
-
-if __name__ == "__main__":
-    main()
